created base canvas class and removed unused methods
Created a base canvas class containing canvas and system related methods that may be used in the future in template builder.
create template area now extends base canvas class
removed a number of methods that are no longer called in the template builder
added comments to most methods
