+++ DataSummitSolution/DataSummitUI/src/app/profileVersion/create/components/createProfileVersion.component.html
@@ -3,13 +3,6 @@
     <input type="file" class="ui-button-rounded ui-button ui-widget ui-state-default ui-corner-all" accept="image/x-png,image/gif,image/jpeg" (change)="readUrl($event);">
     <br />
     <br />
-    <div class="btn-group btn-group-justified" role="group" aria-label="...">
-        <!-- <div class="btn-group" role="group">
-            <button type="button" class="btn btn-outline-danger btn-sm" (click)="removeWhite(url);">
-                <i class="fa fa-times" aria-hidden="true"></i> Remove</button>
-        </div> -->
-    </div>
-    <br />
     <div class="btn-group btn-group-vertical" role="group" aria-label="...">
         <div class="ui-md-6 ui-g-6">
             <button pButton icon="pi pi-refresh" class="ui-button-rounded" pTooltip="Refresh Background" (click)="setCanvasImage(url)"></button>
