+++ DataSummitSolution/DataSummitUI/src/app/profileVersion/create/components/createProfileVersion.component.ts
@@ -11,6 +11,7 @@ import { ProfileAttribute } from '../../../profileAttributes/models/profileAttri
 import { Rectangle } from '../models/rectangle';
 
 import { StandardAttribute } from 'src/app/standardAttribute/models/standardAttribute';
+import { BaseCanvasMethods } from '../models/baseCanvasMethods';
 
 declare const fabric: any;
 
@@ -20,47 +21,27 @@ declare const fabric: any;
   templateUrl: 'createProfileVersion.component.html'
 })
 
-export class CreateProfileVersionComponent implements OnInit, AfterViewInit {
+export class CreateProfileVersionComponent extends BaseCanvasMethods implements OnInit, AfterViewInit {
 
+  @Input() companyId: number;
+  
   @ViewChild('canvasContainer', { static: false }) canvasContainer
   @ViewChild('templateNameModal', { static: false }) templateNameModal
 
-  @Input() companyId: number;
+
   template: ProfileVersion = new ProfileVersion();;
   pairs: RectanglePairs;
   templateName: string;
   loading: boolean;
   standardAttributes: StandardAttribute[] = [];
   headers: string[];
-  canvas: any;
-  
-  props: any = {
-    canvasFill: '#ffffff',
-    canvasImage: '',
-    id: null,
-    opacity: null,
-    fill: null,
-    fontSize: null,
-    lineHeight: null,
-    charSpacing: null,
-    fontWeight: null,
-    fontStyle: null,
-    textAlign: null,
-    fontFamily: null,
-    TextDecoration: ''
-  };
 
+  imageString: string
   textString: string;
   url: any = '';
 
-  json: any;
   globalEditor: boolean = false;
-  textEditor: boolean = false;
-  imageEditor: boolean = false;
-  figureEditor: boolean = false;
-  selected: any;
-  width: number;
-  height: number;
+
   toolbarLeft: number | string = 0;
   toolbarRight: number | string = 'auto';
   toolbarAlign: string;
@@ -72,26 +53,36 @@ export class CreateProfileVersionComponent implements OnInit, AfterViewInit {
   constructor(private router: Router,
     private api: ApiService,
     private route: ActivatedRoute,
-    private notifyService: NotifyService) { }
+    private notifyService: NotifyService) {
+    super()
+  }
 
   ngOnInit() {
     //CompanyId
     this.companyId = this.route.snapshot.params['companyId']
     if (typeof this.companyId == 'string')         //Ensure companyId is number if received as a string
-    { this.companyId = Number(this.companyId);}
+    { this.companyId = Number(this.companyId); }
     this.getStandardAttributeNames();
     this.initProfileVersionTable();
   }
 
+  ngAfterViewInit() {
+    super.initCanvas()
+    this.width = this.canvasContainer.nativeElement.clientWidth - 30
+    this.height = this.width * (1 / 1.4142)
+    this.canvas.setWidth(this.width);
+    this.canvas.setHeight(this.height);
+  }
+
   initProfileVersionTable() {
     this.headers = [
-        "Original",
-        "Standard",
-        "Actions"
-    ]}
+      "Original",
+      "Standard",
+      "Actions"
+    ]
+  }
 
-  saveTemplateName()
-  {
+  saveTemplateName() {
     //Populate profile attributes
     let pas: ProfileAttribute[] = [];
     this.saveGroups.forEach(rp => {
@@ -111,18 +102,19 @@ export class CreateProfileVersionComponent implements OnInit, AfterViewInit {
       pa.ValueY = Math.round(rV.Tl.y);
       if (pa.ProfileVersionId == null) pa.ProfileVersionId = 0;
       if (pa.BlockPositionId == null || pa.BlockPositionId == 0) pa.BlockPositionId = 1;
-      if (pa.StandardAttributeId == null|| pa.StandardAttributeId == 0) pa.StandardAttributeId = 1;
+      if (pa.StandardAttributeId == null || pa.StandardAttributeId == 0) pa.StandardAttributeId = 1;
       pas.push(pa);
     });
     this.template.ProfileAttributes = pas;
     this.templateNameModal.show();
   }
 
-  deleteAttribute(pa: ProfileAttribute)
-  {
+  deleteAttribute(pa: ProfileAttribute) {
     this.template.ProfileAttributes.splice(pa.ProfileAttributeId, 1);
   }
 
+
+  //saves the template
   saveDefinition(templateName: string) {
     this.loading = true;
     this.templateNameModal.hide();
@@ -136,7 +128,7 @@ export class CreateProfileVersionComponent implements OnInit, AfterViewInit {
     let yMax = 0;
     let xMin = 10000000000000;
     let yMin = 10000000000000;
-    
+
     this.template.ProfileAttributes.forEach(attr => {
       attr.StandardAttributeId = parseInt(attr.StandardAttributeId.toString())
       if ((attr.NameX + attr.NameWidth) > xMax) xMax = attr.NameX + attr.NameWidth;
@@ -144,7 +136,7 @@ export class CreateProfileVersionComponent implements OnInit, AfterViewInit {
       if (attr.NameX < xMin) xMin = attr.NameX;
       if (attr.NameY < yMin) yMin = attr.NameY;
     });
-    
+
     this.template.Width = xMax - xMin;
     this.template.Height = yMax - yMin;
     this.template.X = xMin;
@@ -158,19 +150,19 @@ export class CreateProfileVersionComponent implements OnInit, AfterViewInit {
     this.template.ImageString = this.url;
     //TODO replace default userId with actual user
     this.template.UserId = 1;
-    
+
     //let v: string = JSON.stringify(this.template);
 
     this.api.post('api/profileversions', this.template)
-        .pipe(take(1))
-        .subscribe(() => {
-                this.router.navigate([`companies/${this.companyId}/profileversions`])
-                console.log(location.origin.toString() + this.router.url.toString());
-                this.loading = false;
-            }, error => {
-                console.log(error);
-                this.loading = false;
-            })
+      .pipe(take(1))
+      .subscribe(() => {
+        this.router.navigate([`companies/${this.companyId}/profileversions`])
+        console.log(location.origin.toString() + this.router.url.toString());
+        this.loading = false;
+      }, error => {
+        console.log(error);
+        this.loading = false;
+      })
   }
 
   getStandardAttributeNames() {
@@ -193,83 +185,13 @@ export class CreateProfileVersionComponent implements OnInit, AfterViewInit {
     this.standardAttributes.push(new StandardAttribute(17, "Other"));
     this.standardAttributes.push(new StandardAttribute(101, "Drafting Company"));
     this.standardAttributes.push(new StandardAttribute(201, "Client"));
-}
-
-  ngAfterViewInit() {
-    this.canvas = new fabric.Canvas('canvas', {
-      hoverCursor: 'pointer',
-      selection: true,
-      selectionBorderColor: 'blue'
-    });
-
-    this.canvas.on({
-      'object:moving': () => { },
-      'object:modified': () => { },
-      'object:selected': (e) => {
-
-        let selectedObject = e.target;
-        this.selected = selectedObject
-        selectedObject.hasRotatingPoint = true;
-        selectedObject.transparentCorners = false;
-        // selectedObject.cornerColor = 'rgba(255, 87, 34, 0.7)';
-
-        this.resetPanels();
-
-        if (selectedObject.type !== 'group' && selectedObject) {
-
-          this.getId();
-          this.getOpacity();
-
-          switch (selectedObject.type) {
-            case 'rect':
-            case 'circle':
-            case 'triangle':
-              this.figureEditor = true;
-              this.getFill();
-              break;
-            case 'i-text':
-              this.textEditor = true;
-              this.getLineHeight();
-              this.getCharSpacing();
-              this.getBold();
-              this.getFontStyle();
-              this.getFill();
-              this.getTextDecoration();
-              this.getTextAlign();
-              this.getFontFamily();
-              break;
-            case 'image':
-              console.log('image');
-              break;
-          }
-        }
-      },
-      'selection:cleared': () => {
-        this.selected = null;
-        this.resetPanels();
-      }
-    });
-
-    // this.canvas.setWidth(this.size.width);
-    this.width = this.canvasContainer.nativeElement.clientWidth - 30
-    this.height = this.width * (1 / 1.4142)
-    this.canvas.setWidth(this.width);
-    this.canvas.setHeight(this.height);
-
-    // get references to the html canvas element & its context
-    // this.canvas.on('mouse:down', (e) => {
-    // let canvasElement: any = document.getElementById('canvas');
-    // console.log(canvasElement)
-    // });
   }
-
-
-  imageString: string
 
   onUpload() {
 
   }
 
+  //Creates group of the selected rectangles
   linkRects() {
     if (!this.canvas.getActiveObject()) {
       return;
@@ -283,8 +205,7 @@ export class CreateProfileVersionComponent implements OnInit, AfterViewInit {
     if (selectedRects.length == 2) {
       let group: RectanglePair = new RectanglePair();
       selectedRects.forEach(rect => {
-        if (rect.fill == this.titleFill)
-        {
+        if (rect.fill == this.titleFill) {
           let titleRect: Rectangle = new Rectangle();
           titleRect.Tl = rect.aCoords.tl;
           titleRect.Tr = rect.aCoords.tr;
@@ -294,21 +215,20 @@ export class CreateProfileVersionComponent implements OnInit, AfterViewInit {
           titleRect.Height = titleRect.Br.y - titleRect.Tr.y;
           titleRect.Fill = rect.fill;
           titleRect.Type = "title";
-          
+
           group.TitleRectangle = titleRect;
-        } 
-        else if (rect.fill == this.valueFill) 
-        {
+        }
+        else if (rect.fill == this.valueFill) {
           let valueRect: Rectangle = new Rectangle();
           valueRect.Tl = rect.aCoords.tl,
-          valueRect.Tr = rect.aCoords.tr,
-          valueRect.Bl = rect.aCoords.bl,
-          valueRect.Br = rect.aCoords.br,
-          valueRect.Width = valueRect.Tr.x - valueRect.Tl.x,
-          valueRect.Height = valueRect.Br.y - valueRect.Tr.y,
-          valueRect.Fill = rect.fill,
-          valueRect.Type = "value"
-          
+            valueRect.Tr = rect.aCoords.tr,
+            valueRect.Bl = rect.aCoords.bl,
+            valueRect.Br = rect.aCoords.br,
+            valueRect.Width = valueRect.Tr.x - valueRect.Tl.x,
+            valueRect.Height = valueRect.Br.y - valueRect.Tr.y,
+            valueRect.Fill = rect.fill,
+            valueRect.Type = "value"
+
           group.ValueRectangle = valueRect;
         }
       });
@@ -357,6 +277,7 @@ export class CreateProfileVersionComponent implements OnInit, AfterViewInit {
     // }
   }
 
+  //ungroups a selected grouped rectangle pair
   unlinkRects() {
     if (!this.canvas.getActiveObject()) {
       return;
@@ -368,6 +289,7 @@ export class CreateProfileVersionComponent implements OnInit, AfterViewInit {
     this.canvas.requestRenderAll();
   }
 
+  //for testing purposes
   log(e) {
     console.log(e)
   }
@@ -381,79 +303,9 @@ export class CreateProfileVersionComponent implements OnInit, AfterViewInit {
     }
   }
 
-  //Block "Size"
-
-  // changeSize(event: any) {
-  //   this.canvas.setWidth(this.size.width);
-  //   this.canvas.setHeight(this.size.height);
-  // }
-
-  //Block "Add text"
-
-  addText() {
-    let textString = this.textString;
-    let text = new fabric.IText(textString, {
-      left: 10,
-      top: 10,
-      fontFamily: 'helvetica',
-      angle: 0,
-      fill: '#000000',
-      scaleX: 0.5,
-      scaleY: 0.5,
-      fontWeight: '',
-      hasRotatingPoint: true
-    });
-    this.extend(text, this.randomId());
-    this.canvas.add(text);
-    this.selectItemAfterAdded(text);
-    this.textString = '';
-  }
-
-  //Block "Add images"
-
-  getImgPolaroid(event: any) {
-    let el = event.target;
-    fabric.Image.fromURL(el.src, (image) => {
-      image.set({
-        left: 10,
-        top: 10,
-        angle: 0,
-        padding: 10,
-        cornersize: 10,
-        hasRotatingPoint: true,
-        peloas: 12
-      });
-      image.setWidth(150);
-      image.setHeight(150);
-      this.extend(image, this.randomId());
-      this.canvas.add(image);
-      this.selectItemAfterAdded(image);
-    });
-  }
-
   //Block "Upload Image"
 
-  addImageOnCanvas(url) {
-    if (url) {
-      fabric.Image.fromURL(url, (image) => {
-        image.set({
-          left: 10,
-          top: 10,
-          angle: 0,
-          padding: 10,
-          cornersize: 10,
-          hasRotatingPoint: true,
-          width: this.canvasContainer.nativeElement.clientWidth,
-          height: 800
-        });
-        // image.setWidth(200);
-        // image.setHeight(200);
-        this.extend(image, this.randomId());
-        this.canvas.setBackgroundImage(image);
-      });
-    }
-  }
-
+  //Called on selecting an image, sets the canvas background to the selected image and resizes.
   readUrl(event) {
     if (event.target.files && event.target.files[0]) {
       var reader = new FileReader();
@@ -466,25 +318,15 @@ export class CreateProfileVersionComponent implements OnInit, AfterViewInit {
         this.height = image.height
         this.width = image.width
         this.canvas.setBackgroundColor({ source: image.src, repeat: 'no-repeat' }, () => {
-          // self.props.canvasFill = '';
           this.canvas.renderAll();
         });
         this.canvas.renderAll();
-        // this.setCanvasImage(event.target['result']);
       }
       reader.readAsDataURL(event.target.files[0]);
-      //this.ConvertImageToString(reader.result);
     }
   }
 
-  // async ConvertImageToString(im: any)
-  // {
-  //   if (typeof im != "string")
-  //   {
-  //     ArrayBuffer ab = 
-  //   }
-  // }
-
+  //Allows user to move the screen to top/bottom/left/right extremes without the need for scrolling
   panScreen(direction: string) {
     switch (direction) {
       case 'left': window.scrollTo({ left: 0, behavior: "smooth" })
@@ -500,6 +342,7 @@ export class CreateProfileVersionComponent implements OnInit, AfterViewInit {
     }
   }
 
+  //swaps the toolbar orientations
   toggleToolbarSide() {
     if (this.toolbarLeft == 0) {
       this.toolbarRight = 0
@@ -512,7 +355,8 @@ export class CreateProfileVersionComponent implements OnInit, AfterViewInit {
     }
   }
 
-  removeWhite() {
+  //removes image from background.
+  removeImage() {
     this.url = '';
     this.canvas.setBackgroundColor({ source: this.url, repeat: 'no-repeat' }, () => {
       // self.props.canvasFill = '';
@@ -520,22 +364,13 @@ export class CreateProfileVersionComponent implements OnInit, AfterViewInit {
     });
   };
 
-  //Block "Add figure"
-
+  //Adds rectangle to the canvas (title / value rect.)
   addFigure(figure) {
     let add: any;
     console.log(window)
     let x = window.pageXOffset + (window.innerWidth / 2)
     let y = window.pageYOffset + + (window.innerHeight / 2)
     switch (figure) {
-      case 'rectangle':
-        add = new fabric.Rect({
-          width: 200, height: 100, left: 10, top: 10, angle: 0,
-          fill: 'rgba(0,0,0,0.25)',
-          borderColor: '#3f51b5',
-          hasBorders: true
-        });
-        break;
       case 'title':
         add = new fabric.Rect({
           width: 200, height: 100, left: x, top: y, angle: 0,
@@ -552,54 +387,10 @@ export class CreateProfileVersionComponent implements OnInit, AfterViewInit {
           hasBorders: true
         });
         break;
-      case 'square':
-        add = new fabric.Rect({
-          width: 100, height: 100, left: 10, top: 10, angle: 0,
-          fill: '#4caf50'
-        });
-        break;
-      case 'triangle':
-        add = new fabric.Triangle({
-          width: 100, height: 100, left: 10, top: 10, fill: '#2196f3'
-        });
-        break;
-      case 'circle':
-        add = new fabric.Circle({
-          radius: 50, left: 10, top: 10, fill: '#ff5722'
-        });
-        break;
     }
     this.extend(add, this.randomId());
     this.canvas.add(add);
     this.selectItemAfterAdded(add);
-  }
-
-  /*Canvas*/
-
-  cleanSelect() {
-    // this.canvas.deactivateAllWithDispatch().renderAll();
-  }
-
-  selectItemAfterAdded(obj) {
-    // this.canvas.deactivateAllWithDispatch().renderAll();
-    this.canvas.setActiveObject(obj);
-  }
-
-  setCanvasFill() {
-    if (!this.props.canvasImage) {
-      this.canvas.backgroundColor = this.props.canvasFill;
-      this.canvas.renderAll();
-    }
-  }
-
-  extend(obj, id) {
-    obj.toObject = (function (toObject) {
-      return function () {
-        return fabric.util.object.extend(toObject.call(this), {
-          id: id
-        });
-      };
-    })(obj.toObject);
   }
 
   setCanvasImage(url) {
@@ -615,320 +406,5 @@ export class CreateProfileVersionComponent implements OnInit, AfterViewInit {
       this.width = image.width;
       this.height = image.height;
     }
-  }
-
-  randomId() {
-    return Math.floor(Math.random() * 999999) + 1;
-  }
-
-  /*------------------------Global actions for element------------------------*/
-
-  getActiveStyle(styleName, object) {
-    object = object || this.canvas.getActiveObject();
-    if (!object) return '';
-
-    return (object.getSelectionStyles && object.isEditing)
-      ? (object.getSelectionStyles()[styleName] || '')
-      : (object[styleName] || '');
-  }
-
-
-  setActiveStyle(styleName, value, object) {
-    object = object || this.canvas.getActiveObject();
-    if (!object) return;
-
-    if (object.setSelectionStyles && object.isEditing) {
-      var style = {};
-      style[styleName] = value;
-      object.setSelectionStyles(style);
-      object.setCoords();
-    }
-    else {
-      object.set(styleName, value);
-    }
-
-    object.setCoords();
-    this.canvas.renderAll();
-  }
-
-
-  getActiveProp(name) {
-    var object = this.canvas.getActiveObject();
-    if (!object) return '';
-
-    return object[name] || '';
-  }
-
-  setActiveProp(name, value) {
-    var object = this.canvas.getActiveObject();
-    if (!object) return;
-    object.set(name, value).setCoords();
-    this.canvas.renderAll();
-  }
-
-  clone() {
-    let activeObject = this.canvas.getActiveObject();
-
-    if (activeObject) {
-      let clone;
-      switch (activeObject.type) {
-        case 'rect':
-          clone = new fabric.Rect(activeObject.toObject());
-          break;
-        case 'circle':
-          clone = new fabric.Circle(activeObject.toObject());
-          break;
-        case 'triangle':
-          clone = new fabric.Triangle(activeObject.toObject());
-          break;
-        case 'i-text':
-          clone = new fabric.IText('', activeObject.toObject());
-          break;
-        case 'image':
-          clone = fabric.util.object.clone(activeObject);
-          break;
-      }
-      if (clone) {
-        clone.set({ left: 10, top: 10 });
-        this.canvas.add(clone);
-        this.selectItemAfterAdded(clone);
-      }
-    }
-  }
-
-  getId() {
-    this.props.id = this.canvas.getActiveObject().toObject().id;
-  }
-
-  setId() {
-    let val = this.props.id;
-    let complete = this.canvas.getActiveObject().toObject();
-    console.log(complete);
-    this.canvas.getActiveObject().toObject = () => {
-      complete.id = val;
-      return complete;
-    };
-  }
-
-  getOpacity() {
-    this.props.opacity = this.getActiveStyle('opacity', null) * 100;
-  }
-
-  setOpacity() {
-    this.setActiveStyle('opacity', parseInt(this.props.opacity) / 100, null);
-  }
-
-  getFill() {
-    this.props.fill = this.getActiveStyle('fill', null);
-  }
-
-  setFill() {
-    this.setActiveStyle('fill', this.props.fill, null);
-  }
-
-  getLineHeight() {
-    this.props.lineHeight = this.getActiveStyle('lineHeight', null);
-  }
-
-  setLineHeight() {
-    this.setActiveStyle('lineHeight', parseFloat(this.props.lineHeight), null);
-  }
-
-  getCharSpacing() {
-    this.props.charSpacing = this.getActiveStyle('charSpacing', null);
-  }
-
-  setCharSpacing() {
-    this.setActiveStyle('charSpacing', this.props.charSpacing, null);
-  }
-
-  getFontSize() {
-    this.props.fontSize = this.getActiveStyle('fontSize', null);
-  }
-
-  setFontSize() {
-    this.setActiveStyle('fontSize', parseInt(this.props.fontSize), null);
-  }
-
-  getBold() {
-    this.props.fontWeight = this.getActiveStyle('fontWeight', null);
-  }
-
-  setBold() {
-    this.props.fontWeight = !this.props.fontWeight;
-    this.setActiveStyle('fontWeight', this.props.fontWeight ? 'bold' : '', null);
-  }
-
-  getFontStyle() {
-    this.props.fontStyle = this.getActiveStyle('fontStyle', null);
-  }
-
-  setFontStyle() {
-    this.props.fontStyle = !this.props.fontStyle;
-    this.setActiveStyle('fontStyle', this.props.fontStyle ? 'italic' : '', null);
-  }
-
-
-  getTextDecoration() {
-    this.props.TextDecoration = this.getActiveStyle('textDecoration', null);
-  }
-
-  setTextDecoration(value) {
-    let iclass = this.props.TextDecoration;
-    if (iclass.includes(value)) {
-      iclass = iclass.replace(RegExp(value, "g"), "");
-    } else {
-      iclass += ` ${value}`
-    }
-    this.props.TextDecoration = iclass;
-    this.setActiveStyle('textDecoration', this.props.TextDecoration, null);
-  }
-
-  hasTextDecoration(value) {
-    return this.props.TextDecoration.includes(value);
-  }
-
-
-  getTextAlign() {
-    this.props.textAlign = this.getActiveProp('textAlign');
-  }
-
-  setTextAlign(value) {
-    this.props.textAlign = value;
-    this.setActiveProp('textAlign', this.props.textAlign);
-  }
-
-  getFontFamily() {
-    this.props.fontFamily = this.getActiveProp('fontFamily');
-  }
-
-  setFontFamily() {
-    this.setActiveProp('fontFamily', this.props.fontFamily);
-  }
-
-  /*System*/
-
-
-  removeSelected() {
-    let activeObject = this.canvas.getActiveObject()
-    // activeGroup = this.canvas.getActiveGroup();
-
-    if (activeObject) {
-      this.canvas.remove(activeObject);
-      // this.textString = '';
-    }
-    // else if (activeGroup) {
-    //   let objectsInGroup = activeGroup.getObjects();
-    //   this.canvas.discardActiveGroup();
-    //   let self = this;
-    //   objectsInGroup.forEach(function (object) {
-    //     self.canvas.remove(object);
-    //   });
-    // }
-  }
-
-  bringToFront() {
-    let activeObject = this.canvas.getActiveObject(),
-      activeGroup = this.canvas.getActiveGroup();
-
-    if (activeObject) {
-      activeObject.bringToFront();
-      // activeObject.opacity = 1;
-    }
-    else if (activeGroup) {
-      let objectsInGroup = activeGroup.getObjects();
-      this.canvas.discardActiveGroup();
-      objectsInGroup.forEach((object) => {
-        object.bringToFront();
-      });
-    }
-  }
-
-  sendToBack() {
-    let activeObject = this.canvas.getActiveObject(),
-      activeGroup = this.canvas.getActiveGroup();
-
-    if (activeObject) {
-      activeObject.sendToBack();
-      // activeObject.opacity = 1;
-    }
-    else if (activeGroup) {
-      let objectsInGroup = activeGroup.getObjects();
-      this.canvas.discardActiveGroup();
-      objectsInGroup.forEach((object) => {
-        object.sendToBack();
-      });
-    }
-  }
-
-  confirmClear() {
-    if (confirm('Are you sure?')) {
-      this.canvas.clear();
-    }
-  }
-
-  rasterize() {
-    if (!fabric.Canvas.supports('toDataURL')) {
-      alert('This browser doesn\'t provide means to serialize canvas to an image');
-    }
-    else {
-      console.log(this.canvas.toDataURL('png'))
-      //window.open(this.canvas.toDataURL('png'));
-      var image = new Image();
-      image.src = this.canvas.toDataURL('png')
-      var w = window.open("");
-      w.document.write(image.outerHTML);
-    }
-  }
-
-  rasterizeSVG() {
-    console.log(this.canvas.toSVG())
-    // window.open(
-    //   'data:image/svg+xml;utf8,' +
-    //   encodeURIComponent(this.canvas.toSVG()));
-    // console.log(this.canvas.toSVG())
-    // var image = new Image();
-    // image.src = this.canvas.toSVG()
-    var w = window.open("");
-    w.document.write(this.canvas.toSVG());
-  };
-
-
-  saveCanvasToJSON() {
-    let json = JSON.stringify(this.canvas);
-    localStorage.setItem('Kanvas', json);
-    console.log('json');
-    console.log(json);
-
-  }
-
-  loadCanvasFromJSON() {
-    let CANVAS = localStorage.getItem('Kanvas');
-    console.log('CANVAS');
-    console.log(CANVAS);
-
-    // and load everything from the same json
-    this.canvas.loadFromJSON(CANVAS, () => {
-      console.log('CANVAS untar');
-      console.log(CANVAS);
-
-      // making sure to render canvas at the end
-      this.canvas.renderAll();
-
-      // and checking if object's "name" is preserved
-      console.log('this.canvas.item(0).name');
-      console.log(this.canvas);
-    });
-
-  };
-
-  rasterizeJSON() {
-    this.json = JSON.stringify(this.canvas, null, 2);
-  }
-
-  resetPanels() {
-    this.textEditor = false;
-    this.imageEditor = false;
-    this.figureEditor = false;
   }
 }
